// frontend/src/lib/api/model/nodeRole.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type NodeRole = (typeof NodeRole)[keyof typeof NodeRole];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NodeRole = {
  MASTER: "MASTER",
  REPLICA: "REPLICA",
  PROXY: "PROXY",
  ORCHESTRATOR: "ORCHESTRATOR",
} as const;