// frontend/src/lib/api/model/nodeStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type NodeStatus = (typeof NodeStatus)[keyof typeof NodeStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NodeStatus = {
  STARTING: "STARTING",
  RUNNING: "RUNNING",
  STOPPED: "STOPPED",
  FAILED: "FAILED",
  SYNCING: "SYNCING",
} as const;