// frontend/src/lib/api/model/nodeStatsResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface NodeStatsResponse {
  nodeId?: string;
  containerName?: string;
  nodeRole?: string;
  timestamp?: string;
  cpuUsagePercent?: number;
  memoryUsage?: number;
  memoryLimit?: number;
  memoryUsagePercent?: number;
  networkRxBytes?: number;
  networkTxBytes?: number;
  blockRead?: number;
  blockWrite?: number;
  running?: boolean;
  state?: string;
  replicationLagSeconds?: number;
}